Repository Refactoring + Tests Updated

+++ VenomPizzaMenuService/src/repository/ProductsRepository.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Storage;
 using System.Runtime.InteropServices;
 using System.Threading.Tasks;
 using VenomPizzaMenuService.src.context;
@@ -22,9 +23,6 @@ public class ProductsRepository : IProductsRepository
         try
         {
             var addedProduct = await AddProductToDatabase(newProduct);
-            var priceVariants = newProduct.PriceVariantsDict.Select((x, y) => new PriceVariant(addedProduct, y, x.Key, x.Value));
-            _dbContext.PriceVariants.AddRange(priceVariants);
-            await _dbContext.SaveChangesAsync();
             await transaction.CommitAsync();
             return addedProduct;
         }
@@ -40,19 +38,8 @@ public class ProductsRepository : IProductsRepository
         await using var transaction=await _dbContext.Database.BeginTransactionAsync();
         try
         {
-            var idList = newCombo.ProductsDict.Keys.ToList();
-            var foundedProducts = await _dbContext.Products.Where(x => idList.Contains(x.Id)).ToListAsync();
-            foreach (var searchedProduct in newCombo.ProductsDict)
-                if (!foundedProducts.Any(x => x.Id == searchedProduct.Key))
-                    throw new ArgumentException("Товара с ID " + searchedProduct.Key + " не существует");
-
             var addedCombo = await AddProductToDatabase(newCombo);
             _dbContext.ComboProducts.AddRange(newCombo.ProductsDict.Select(x => new ComboProduct(addedCombo.Id, x.Key, x.Value)));
-
-            var priceVariants = newCombo.PriceVariantsDict.Select((x, y) => new PriceVariant(addedCombo, y, x.Key, x.Value));
-            _dbContext.PriceVariants.AddRange(priceVariants);
-
-            await _dbContext.SaveChangesAsync();
             await transaction.CommitAsync();
             return addedCombo;
         }
@@ -90,7 +77,7 @@ public class ProductsRepository : IProductsRepository
         return result;
     }
 
-    public async Task<Product> GetProductById(int id)
+    public async Task<Product?> GetProductById(int id)
     {
         var foundedProduct= await _dbContext.Products
             .Include(p=>(p as Combo).Products)
@@ -98,22 +85,31 @@ public class ProductsRepository : IProductsRepository
             .Include(p=>p.PriceVariants)
             .AsSplitQuery()
             .FirstOrDefaultAsync(x => x.Id == id);
-        if (foundedProduct == null)
-            throw new KeyNotFoundException("Товара с ID " + id + " не найдено");
         return foundedProduct;
     }
 
-    public async Task<(int id,string title)> GetProductIdAndTitle(int productId,int priceId)
+    public async Task<(int id,string title)?> GetProductIdAndTitle(int productId,int priceId)
     {
         var foundedProduct=await _dbContext.Products
             .AsNoTracking()
             .Include(x=>x.PriceVariants)
             .FirstOrDefaultAsync(x=>x.Id== productId);
-        if (foundedProduct == null)
-            throw new KeyNotFoundException($"Товар с Id {productId} не найден");
-        if (foundedProduct.PriceVariants.FirstOrDefault(x => x.PriceId == priceId) == null)
-            throw new KeyNotFoundException($"Цена {priceId} товара {productId} не найдена");
+        if (foundedProduct == null || foundedProduct.PriceVariants.FirstOrDefault(x => x.PriceId == priceId) == null)
+            return null;
         return (foundedProduct.Id, foundedProduct.Title);
+    }
+
+    public async Task CheckComboProducts(ComboDto newCombo)
+    {
+        var idList = newCombo.ProductsDict.Keys.ToList();
+        var foundedProducts = await _dbContext.Products.Where(x => idList.Contains(x.Id)).ToListAsync();
+        foreach (var searchedProduct in newCombo.ProductsDict)
+        {
+            if (!foundedProducts.Any(x => x.Id == searchedProduct.Key))
+                throw new KeyNotFoundException($"Товара с ID {searchedProduct.Key} не существует");
+            if (foundedProducts.First(x => x.Id == searchedProduct.Key) is Combo)
+                throw new ArgumentException($"Нельзя добавить комбо с ID {searchedProduct.Key} в новое комбо");
+        }
     }
     #endregion
 
@@ -127,17 +123,7 @@ public class ProductsRepository : IProductsRepository
             if (foundedProduct == null)
                 throw new KeyNotFoundException("Товара с ID " + updatedProduct.Id + " не найдено");
 
-            _dbContext.Entry(foundedProduct).CurrentValues.SetValues(updatedProduct);
-            _dbContext.PriceVariants.RemoveRange(foundedProduct.PriceVariants);
-
-            var priceVariants = updatedProduct.PriceVariantsDict.Select((x, y) => new PriceVariant(foundedProduct, y, x.Key, x.Value));
-            _dbContext.PriceVariants.AddRange(priceVariants);
-
-            if (foundedProduct is Dish foundedDish && updatedProduct is DishDto updatedDish)
-                UpdateDish(foundedDish, updatedDish);
-            else if (foundedProduct is Combo foundedCombo && updatedProduct is ComboDto updatedCombo)
-                UpdateCombo(foundedCombo, updatedCombo);
-
+            UpdateProduct(foundedProduct, updatedProduct);
             await _dbContext.SaveChangesAsync();
             await transaction.CommitAsync();
             return foundedProduct;
@@ -151,38 +137,61 @@ public class ProductsRepository : IProductsRepository
     #endregion
 
     #region delete
-    public async Task DeleteProductById(int id)
+    public async Task<bool> DeleteProductById(int id)
     {
         var foundedProduct = await GetProductById(id);
+        if (foundedProduct == null)
+            return false;
         _dbContext.Products.Remove(foundedProduct);
         await _dbContext.SaveChangesAsync();
+        return true;
     }
     #endregion
 
     #region private
-    private void UpdateDish(Dish subject, DishDto dto)
+    private void UpdateProduct(Product subject, ProductDto newObject)
     {
-        subject.Proteins = dto.Proteins;
-        subject.Fats = dto.Fats;
-        subject.Carbohydrates = dto.Carbohydrates;
-        subject.Calorific = dto.Calorific;
-        subject.Ingredients = dto.Ingredients;
-        subject.Allergens = dto.Allergens;
+        _dbContext.Entry(subject).CurrentValues.SetValues(newObject);
+        _dbContext.PriceVariants.RemoveRange(subject.PriceVariants);
+
+        var priceVariants = newObject.PriceVariantsDict.Select((x, y) => new PriceVariant(subject, y, x.Key, x.Value));
+        _dbContext.PriceVariants.AddRange(priceVariants);
+
+        if (subject is Dish foundedDish && newObject is DishDto updatedDish)
+            UpdateDishFields(foundedDish, updatedDish);
+        else if (subject is Combo foundedCombo && newObject is ComboDto updatedCombo)
+            RecreateComboProducts(foundedCombo, updatedCombo);
     }
-    private async Task UpdateCombo(Combo subject, ComboDto dto)
+
+    private void UpdateDishFields(Dish subject, DishDto newObject)
+    {
+        subject.Proteins = newObject.Proteins;
+        subject.Fats = newObject.Fats;
+        subject.Carbohydrates = newObject.Carbohydrates;
+        subject.Calorific = newObject.Calorific;
+        subject.Ingredients = newObject.Ingredients;
+        subject.Allergens = newObject.Allergens;
+    }
+
+    private void RecreateComboProducts(Combo subject, ComboDto dto)
     {
         if (dto.Products == null)
             return;
         _dbContext.ComboProducts.RemoveRange(subject.Products);
         _dbContext.ComboProducts.AddRange(dto.ProductsDict.Select(x => new ComboProduct(subject.Id, x.Key, x.Value)));
     }
+
     private async Task<Product> AddProductToDatabase(ProductDto newProduct)
     {
         if (await _dbContext.Products.AnyAsync(x => x.Id == newProduct.Id))
             throw new ArgumentException("Товар с ID " + newProduct.Id + " уже существует");
-        var addedProduct = _dbContext.Products.Add(newProduct.ToProduct());
+        var addedProduct = _dbContext.Products.Add(newProduct.ToProduct()).Entity;
         await _dbContext.SaveChangesAsync();
-        return addedProduct.Entity;
+
+        var priceVariants = newProduct.PriceVariantsDict.Select((x, y) => new PriceVariant(addedProduct, y, x.Key, x.Value));
+        _dbContext.PriceVariants.AddRange(priceVariants);
+        await _dbContext.SaveChangesAsync();
+        return addedProduct;
     }
     #endregion
 }
